Almost done sign up
I have almost finished the login. Things yet to do are to add
placeholders. It would be nice if when users look at the terms of
service they would have a button to check box.

There is still a lot of work to do if the user enter invalid input there
must be a mechanism to show them. Twitter bootstrap has the styles but I
am not sure how to implement without talking to our controller team.